feat(metrics): Add metrics schema and related database migrations, including metrics_file table and seeding for development

+++ regtech-app/src/main/resources/db/migration/metrics/V1__create_metrics_file_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS metrics.metrics_file (
     filename VARCHAR(255) PRIMARY KEY,
     date VARCHAR(255),
-    score DOUBLE,
+    score DOUBLE PRECISION,
     status VARCHAR(255),
     batch_id VARCHAR(255),
     bank_id VARCHAR(255)
